Refactor the db initialization scaffolding to use swh.core.db
ie. split the 2 provenance.sql scripts in parts and move them to
swh/provenance/sql so the database creation and initialization are handled
by the `swh db init` and `swh db create` tools from swh.core.db.

The with/witout path choice is implemented using 2 flavors: with-path
and without-path. So the initialization command should be something like:

  swh db init -d db-name --flavor with-path provenance