Improve DB schema naming
make it more consistent, so we can then simplify the code a bit by
making relation table names composable (i.e. the relation table between
content and revision is now simply content_in_revision, and so on) as
well as column names (e.g. rev->revision and so on).

Also remove the conditional sql execution in db init code (in
swh/provennce/sql/), because it is actually unnecessary: keeping the
`location` column for the "without-path" flavor does not cost anything.
The only regression is that the location column in entity tables lost
its "not null" constraint, but it's a minor drawback that the gain in
clarity and simplicity of the db initialization code make up.