Use the flavor in the db to choose the ProvenanceDB class to instanciate
instead of a config option that may be inconsistent with how the
database has been created/populated.

So now, once the flavor of the provenance DB is chosen, at db creation
time, no need for specifying the "with_path" config option.

Deprecate the "with_path" config option in get_provenance(), but make
sure the given value, if any, is consistent with the database.

Modify the provenance pytest fixture to provide both versions of the
provenance DB (with and without path).

Adapt tests to make them pass for the (previously untested) provenance DB
wihout path flavor.