Refactor the generate_storage_from_git dataset creation tool
use the newly introduced swh.journal StreamJournalWriter to generate the
msgpack-file with the content of the storage serialized, instead of
reimplementing a similar logic by hand.

Doing so, the msgpack file format is a little bit different, so current
datasets need to be regenerated.

From the 'consumer' side (the fill_storage fixture), simply use
swh.storage's process_replay_objects (once again, instead of reinventing
the wheel).

This is needed because we will need to handle much more model object types
for testing the origin-revision layer.