Improve PostgreSQL storage scheme for the `with-path-denormalized` flavor
Previous version was storing arrays of strings representing tuples for the
denormalized relations (`dst` and `loc` of the relation resp.). While that
simplified the check for duplicates, it turned out to be very inefficient
in terms of disk usage. The new version has two distinct lists if `bigint`
(ie. internal ids) for `dst` and `loc` resp. To check for duplicates the
lists should be zipped, and repeated tuples filtered.