Add explicit flag for flattenned directories to `ProvenanceStorageInterface`
Both contents and directories should always have an associated date in the storage.
Flattening of a directory is now explicitly acknowledged by setting the newly added
flag. The idea is to allow to postpone the creation of flat models for directories
in the isochrone frontier (the algorithm will be refactored in the commits to come).