Limit the number of entries in the cache
The implementation test after each revision if there are
more than 100 000 entries of any kind in the cache. If yes,
it flush the content.
It's quite naive but the memory seems to stay around 5Go
max on big repositories like linux.

This can be still an issue with the current zeromq client because
the origin are sorted by repositories a lot of snapshots of big
repositories can be ingested in parallel.
With the journal implementation, it should be more distributed on
repositories of different sizes

Related to T4313