postgresql archive: add support for partially copied databases
The incremental copy of the archive to mmca is not atomic: the directory
table needs to be copied first, then the directory_entry_* tables need
to be updated. This means that the client can view inconsistent entries,
where the directory has been synced but not all the entry rows.

We return an empty list when one of these bogus entries is detected.
This allows smooth fallback to the main database through the
multiplexer.