provenance: Fix conditional on anchor
The anchor can be an empty string in case nothing is found. So that fails the
conditional to return None currently and then raises when trying to make a
QualifiedSWHID out of an empty string.

Resolves

```
ValidationError: /api/1/provenance/whereis/{target}/
Invalid SWHID: invalid syntax: %(swhid)s
```

Refs. swh/infra/sysadm-environment#5624